Pack structs more efficiently, -Wpadded added to build flags. Fixes #79.
This dropped the interpreter's basic memory use (the amount used to parse
absolutely nothing) down by about 1K. 8800 bytes is bigger than I'd like, but
that around 1K is a big goal for simply shuffling things around.

+++ src/lily_seed_symtab.h
@@ -14,9 +14,9 @@
 /* Sync name order with SYM_CLASS_* #defines in lily_symtab.h */
 typedef const struct {
     char *name;
-    int is_refcounted;
-    int template_count;
-    int flags;
+    uint16_t is_refcounted;
+    uint16_t template_count;
+    uint32_t flags;
     char *parent_name;
     const lily_prop_seed_t *prop_seeds;
     class_setup_func setup_func;
